refactor(guardian): extract _run_subprocess to shared _subprocess module
Move the async subprocess helper from health_signals.py to its own
guardian/_subprocess.py module. It was imported by 5 files (cgroup_ops,
snapshots, recovery, collector, health_signals) — a private function
with public-level usage. Now properly housed with a clean import path.
